move newRepositoryForm component and move tests to bh
