+++ src/components-v1/shared/NewRepositoryModal.vue
@@ -13,7 +13,7 @@
 </template>
 
 <script>
-import NewRepositoryForm from '@/components-v1/repository/NewRepositoryForm';
+import NewRepositoryForm from '@/components/repository/NewRepositoryForm';
 
 const components = {
   NewRepositoryForm,
